requirements.txt
sqlalchemy
pymupdf
python-docx
openpyxl
requests
pytest
python-dotenv
flake8
black
mypy
streamlit>=1.25.0
prometheus-client
psycopg2-binary
pydantic-settings
pydantic
pandas
langchain
uvicorn[standard]
fastapi
faiss-cpu